+++ src/main/java/lfie/danbro/community/community/dto/QuestionDto.java
@@ -6,15 +6,15 @@ import lombok.Data;
 
 @Data
 public class QuestionDto {
-    private Integer id;
+    private Long id;
     private String title;
     private String description;
     private Long gmtCreate;
     private Long gmtModified;
     private Integer creator;
-    private Integer commentCount;
-    private Integer viewCount;
-    private Integer likeCount;
+    private Long commentCount;
+    private Long viewCount;
+    private Long likeCount;
     private String tag;
     private User user;
